PR-005: move public DTOs and MemoryType into src/api/types
