+++ src/api/types/memory_input.rs
@@ -2,7 +2,7 @@ use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::models::memory::MemoryType;
+use super::MemoryType;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct MemoryInput {
