+++ src/api/types/mod.rs
@@ -1,5 +1,11 @@
-// Temporary re-exports during PR-005 migration.
-// Once DTOs are moved into `src/api/types/**`, these will become real module declarations.
+mod memory;
+mod memory_filters;
+mod memory_input;
+mod memory_type;
+mod scored_memory;
 
-pub use crate::models::memory::dto::{Memory, MemoryFilters, MemoryInput, ScoredMemory};
-pub use crate::models::memory::MemoryType;
+pub use memory::Memory;
+pub use memory_filters::MemoryFilters;
+pub use memory_input::MemoryInput;
+pub use memory_type::MemoryType;
+pub use scored_memory::ScoredMemory;
