+++ src/api/types/scored_memory.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::models::memory::dto::Memory;
+use super::Memory;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ScoredMemory {
