+++ src/models/memory/memory_entry.rs
@@ -2,9 +2,8 @@ use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
+use crate::api::types::{Memory, MemoryType};
 use crate::errors::CustomError;
-use crate::models::memory::dto::Memory;
-use crate::models::memory::MemoryType;
 use crate::models::vector::VectorMetadata;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
